add core services to add/delete reviews, as well as remove bootstrap dependencies

+++ anglee-dashboard-ui/src/components/Admin/MovieCards/MovieCard/MovieCard.js
@@ -1,25 +1,44 @@
 import React from "react";
-import { Card, Icon} from "antd";
+import { Card, Icon, Tooltip } from "antd";
+import styled from "styled-components";
 const { Meta } = Card;
 
+const CoverImg = styled.div`
+  height: 200px;
+  background-image: ${props => `url(${props.url})`};
+  background-size: cover;
+`;
+
 const MovieCard = props => {
-  const handleDelete = () => {
-    const movieId = props.movie._id;
-    props.deleteMovie(movieId);
+  const handleDelete = () => props.deleteMovie(props.movie);
+  const handleCrawl = () => props.crawlTheInternet(props.movie);
+  const dateStr = props.movie.releaseDate.split("T")[0];
+
+  let actions = {};
+  if (props.isAdmin) {
+    actions = [
+      <Tooltip placement="bottom" title={"edit"}>
+        <Icon type="edit" />
+      </Tooltip>,
+      <Tooltip placement="bottom" title={"crawl"}>
+        <Icon type="rocket" onClick={handleCrawl} />
+      </Tooltip>,
+      <Tooltip placement="bottom" title={"delete"}>
+        <Icon type="delete" onClick={handleDelete} />
+      </Tooltip>
+    ];
   }
+
   return (
     <Card
       hoverable
-      style={{ width: 300 }}
-      cover={<img src={props.movie.imageUrl} alt={props.movie.title} />}
-      actions={[
-        <Icon type="edit" />,
-        <Icon type="delete" onClick={handleDelete} />
-      ]}
+      // style={{ width: 300 }}
+      cover={<CoverImg url={props.movie.imageUrl} />}
+      actions={actions}
       loading={props.loading}
       size="small"
     >
-      <Meta title={props.movie.title} description={props.movie.description} />
+      <Meta title={props.movie.title} description={dateStr} />
     </Card>
   );
 };
